+++ concordium-node/src/connection/message_handlers.rs
@@ -15,12 +15,14 @@ use crate::{
     read_or_die,
 };
 use anyhow::{bail, ensure};
+use tokio::sync::OwnedSemaphorePermit;
 
 impl Connection {
     /// Processes a network message based on its type.
     pub fn handle_incoming_message(
         &mut self,
         msg: NetworkMessage,
+        permit: OwnedSemaphorePermit,
         conn_stats: &[PeerStats],
     ) -> anyhow::Result<()> {
         // the handshake should be the first incoming network message
@@ -83,7 +85,7 @@ impl Connection {
             }
             NetworkPayload::NetworkPacket(pac, ..) => {
                 // packet receipt is logged later, along with its contents
-                self.handle_incoming_packet(pac, peer_id)
+                self.handle_incoming_packet(pac, permit, peer_id)
             }
         }
     }
@@ -166,6 +168,7 @@ impl Connection {
     fn handle_incoming_packet(
         &self,
         pac: NetworkPacket,
+        permit: OwnedSemaphorePermit,
         peer_id: RemotePeerId,
     ) -> anyhow::Result<()> {
         let is_broadcast = matches!(pac.destination, PacketDestination::Broadcast(..));
@@ -176,6 +179,7 @@ impl Connection {
             vec![peer_id],
             peer_id,
             pac.message,
+            permit,
             is_broadcast,
         )
     }