+++ concordium-node/src/connection/mod.rs
@@ -10,9 +10,12 @@ use bytesize::ByteSize;
 use circular_queue::CircularQueue;
 use low_level::ConnectionLowLevel;
 use mio::{net::TcpStream, Interest, Token};
+use tokio::sync::OwnedSemaphorePermit;
 
+use crate::consensus_ffi::helpers::PacketType;
 #[cfg(feature = "network_dump")]
 use crate::dumper::DumpItem;
+use crate::network::serialization::fbs::PEER_LIST_LIMIT;
 use crate::{
     common::{
         get_current_stamp,
@@ -30,8 +33,6 @@ use crate::{
     read_or_die, write_or_die,
 };
 
-use crate::consensus_ffi::helpers::PacketType;
-
 use std::{
     collections::VecDeque,
     convert::TryFrom,
@@ -44,8 +45,6 @@ use std::{
         Arc, RwLock,
     },
 };
-
-use crate::network::serialization::fbs::PEER_LIST_LIMIT;
 
 /// Designates the sending priority of outgoing messages.
 // If a message is labelled as having `High` priority it is always pushed to the
@@ -386,6 +385,16 @@ pub struct Connection {
     pub pending_messages: MessageQueues,
     /// The wire protocol version for communicating on the connection.
     pub wire_version: WireProtocolVersion,
+    /// Semaphore for pending messages across all incoming messages processing queues.
+    /// When this semaphore reaches 0, new messages from this peer
+    /// will not be read until the pending messages have been dequeued.
+    /// This prevents a single peer from disproportionately filling up the queue.
+    pub pending_messages_semaphore: Arc<tokio::sync::Semaphore>,
+    /// The value will be set to `true` when above semaphore reaches 0,
+    /// allowing the node to re-check at every read polling cycle
+    /// if a semaphore permit got released so remaining incoming messages
+    /// can be read from the peer's socket.
+    pub pending_messages_semaphore_reached: bool,
     /// Semaphore to limit concurrent processing of GetPeers requests.
     pub get_peers_list_semaphore: Arc<tokio::sync::Semaphore>,
 }
@@ -448,6 +457,10 @@ impl Connection {
             // When we create the connection, we set the wire protocol version
             // to the current version, but this is overwritten in the handshake.
             wire_version: WIRE_PROTOCOL_CURRENT_VERSION,
+            pending_messages_semaphore: Arc::new(tokio::sync::Semaphore::new(
+                handler.config.max_queued_messages_per_peer,
+            )),
+            pending_messages_semaphore_reached: false,
             // semaphore starts at 1 to cater for bootstrapper node sending peers list unsolicitedly
             get_peers_list_semaphore: Arc::new(tokio::sync::Semaphore::new(1)),
         })
@@ -527,14 +540,43 @@ impl Connection {
         Ok(is_duplicate)
     }
 
-    /// Keeps reading from the socket as long as there is data to be read
-    /// and the operation is not blocking.
+    /// Keeps reading from the socket as long as there is data to be read, the operation is not blocking
+    /// and the sending peer's `pending_messages_semaphore` value hasn't be exhausted.
     /// The return value indicates if the connection is still open.
     #[inline]
     pub fn read_stream(&mut self, conn_stats: &[PeerStats]) -> anyhow::Result<bool> {
+        // Acquire an owned permit
+        let mut permit = match self.pending_messages_semaphore.clone().try_acquire_owned() {
+            Ok(permit) => permit,
+            Err(_) => {
+                trace!(
+            "Delaying read from peer `{:?}` as it reached its `pending_messages_semaphore` limit",
+            self.remote_peer
+        );
+                self.pending_messages_semaphore_reached = true;
+                return Ok(true);
+            }
+        };
+        self.pending_messages_semaphore_reached = false;
+
         loop {
             match self.low_level.read_from_socket()? {
-                ReadResult::Complete(msg) => self.process_message(Arc::from(msg), conn_stats)?,
+                ReadResult::Complete(msg) => {
+                    self.process_message(Arc::from(msg), permit, conn_stats)?;
+
+                    // Acquire an owned permit again for the next loop cycle
+                    permit = match self.pending_messages_semaphore.clone().try_acquire_owned() {
+                        Ok(permit) => permit,
+                        Err(_) => {
+                            trace!(
+            "Delaying read from peer `{:?}` as it reached its `pending_messages_semaphore` limit",
+            self.remote_peer
+        );
+                            self.pending_messages_semaphore_reached = true;
+                            return Ok(true);
+                        }
+                    };
+                }
                 ReadResult::Incomplete => {}
                 ReadResult::WouldBlock => return Ok(true),
                 ReadResult::Closed => return Ok(false),
@@ -546,6 +588,7 @@ impl Connection {
     fn process_message(
         &mut self,
         bytes: Arc<[u8]>,
+        permit: OwnedSemaphorePermit,
         conn_stats: &[PeerStats],
     ) -> anyhow::Result<()> {
         self.update_last_seen();
@@ -579,7 +622,7 @@ impl Connection {
         }
 
         // process the incoming message
-        self.handle_incoming_message(message, conn_stats)
+        self.handle_incoming_message(message, permit, conn_stats)
     }
 
     /// Concludes the connection's handshake process.