+++ concordium-node/src/p2p/maintenance.rs
@@ -19,7 +19,10 @@ use crate::{
     connection::{ConnChange, Connection, DeduplicationHashAlgorithm, DeduplicationQueues},
     consensus_ffi::{
         catch_up::PeerList,
-        consensus::{ConsensusContainer, Regenesis, CALLBACK_QUEUE},
+        consensus::{
+            ConsensusContainer, Regenesis, CALLBACK_QUEUE, CONSENSUS_QUEUE_DEPTH_IN_BG,
+            CONSENSUS_QUEUE_DEPTH_IN_HI, CONSENSUS_QUEUE_DEPTH_IN_LO,
+        },
     },
     lock_or_die,
     network::{Buckets, NetworkId, Networks},
@@ -56,6 +59,7 @@ use std::{
 pub struct NodeConfig {
     pub no_net: bool,
     pub desired_nodes_count: u16,
+    pub max_queued_messages_per_peer: usize,
     pub no_bootstrap_dns: bool,
     /// Clear persistent bans on startup.
     pub clear_bans: bool,
@@ -214,6 +218,9 @@ pub struct BadEvents {
     /// Number of low priority messages that were dropped because they could not
     /// be enqueued.
     pub dropped_low_queue: Mutex<HashMap<RemotePeerId, u64>>,
+    /// Number of background messages that were dropped because they could not
+    /// be enqueued.
+    pub dropped_background_queue: Mutex<HashMap<RemotePeerId, u64>>,
     /// Number of invalid messages received from the given peer.
     pub invalid_messages: Mutex<HashMap<RemotePeerId, u64>>,
 }
@@ -232,6 +239,15 @@ impl BadEvents {
     /// dropped high priority messages for the peer.
     pub fn inc_dropped_low_queue(&self, peer_id: RemotePeerId) -> u64 {
         *lock_or_die!(self.dropped_low_queue)
+            .entry(peer_id)
+            .and_modify(|x| *x += 1)
+            .or_insert(1)
+    }
+
+    /// Register a new dropped value for the given peer and return the amount of
+    /// dropped background messages for the peer.
+    pub fn inc_dropped_background_queue(&self, peer_id: RemotePeerId) -> u64 {
+        *lock_or_die!(self.dropped_background_queue)
             .entry(peer_id)
             .and_modify(|x| *x += 1)
             .or_insert(1)
@@ -336,9 +352,27 @@ impl P2PNode {
 
         let given_addresses = RwLock::new(parse_config_nodes(&conf.connection)?);
 
+        // Should not overflow as multiplication is between `u16` and `usize` types.
+        let max_queue_size: u128 = (conf.connection.desired_nodes as u128)
+            * conf.connection.max_queued_messages_per_peer as u128;
+
+        let min_queue_size = CONSENSUS_QUEUE_DEPTH_IN_HI
+            .min(CONSENSUS_QUEUE_DEPTH_IN_LO)
+            .min(CONSENSUS_QUEUE_DEPTH_IN_BG);
+
+        if max_queue_size > min_queue_size as u128 {
+            warn!(
+                "The max queue size (desired_nodes * max_queued_messages_per_peer) = {} \
+        exceeds the smallest incoming consensus queue size of {}. \
+        Adjust the relevant node environment variables or configuration 
+        as inbound messages could be dropped under such configuration.",
+                max_queue_size, min_queue_size
+            );
+        }
         let config = NodeConfig {
             no_net: conf.cli.no_network,
             desired_nodes_count: conf.connection.desired_nodes,
+            max_queued_messages_per_peer: conf.connection.max_queued_messages_per_peer,
             no_bootstrap_dns: conf.connection.no_bootstrap_dns,
             clear_bans: conf.connection.clear_bans,
             disallow_multiple_peers_on_ip: conf.connection.disallow_multiple_peers_on_ip,