+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## Unreleased changes
+
 - Prohibit peers from sending unsolicited PeerList messages
+- Enhance node performance by limiting queue saturation from peers that send messages aggressively by using backpressure.
+- Introduce a background queue for processing messages that don't require the global block state lock. 
 
 # 10.0.5